+++ package.json
@@ -8,11 +8,8 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-<<<<<<< HEAD
+    "react-router": "^6.21.2",
     "react-router-dom": "^6.21.2",
-=======
-    "react-paginate": "^8.2.0",
->>>>>>> 4919101612ef7219c355d004caa5be4a89283aa4
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
