+++ src/App.js
@@ -4,8 +4,8 @@ import './Components/footer/Footer.css';
 import Header from './Components/header/Header';
 import './Components/header/Header.css';
 import React, { useState, useEffect } from 'react';
-import Pagination from './Pagination';
-import './pagination.css'
+import Pagination from './Components/pagination/Pagination';
+import './Components/pagination/Pagination.css'
 // import NavigationBar from './NavegationBar';
 // import ProductListFiltered from './ProductListFiltered';
